Removed file not in use Small refactoring to limit repeated code Separated out Entrypoint Create Instructions to separate file
